+++ src/components/Create/components/RewardsList/RewardsList.tsx
@@ -4,7 +4,7 @@ import { useModal } from 'hooks/Modal'
 import { FormItemInput } from 'models/formItemInput'
 import { createContext, useCallback, useContext, useState } from 'react'
 import { MAX_NFT_REWARD_TIERS } from 'utils/nftRewards'
-import { CreateButton } from '../CreateButton'
+import { CreateButton } from 'components/CreateButton'
 import { AddEditRewardModal } from './AddEditRewardModal'
 import { useRewards } from './hooks'
 import { RewardItem } from './RewardItem'
