+++ src/components/Create/components/pages/FundingCycles/FundingCyclesPage.tsx
@@ -12,15 +12,15 @@ import { JuiceDatePicker } from 'components/inputs/JuiceDatePicker'
 import moment from 'moment'
 import { useContext, useEffect } from 'react'
 import { useSetCreateFurthestPageReached } from 'redux/hooks/EditingCreateFurthestPageReached'
+import { durationMustExistRule } from 'utils/antd-rules'
 import { CreateBadge } from '../../CreateBadge'
 import { CreateCollapse } from '../../CreateCollapse'
-import { DurationInput } from '../../DurationInput'
+import { DurationInput } from 'components/DurationInput'
 import { Icons } from '../../Icons'
 import { OptionalHeader } from '../../OptionalHeader'
 import { Selection } from '../../Selection/Selection'
 import { Wizard } from '../../Wizard'
 import { PageContext } from '../../Wizard/contexts/PageContext'
-import { durationMustExistRule } from '../utils'
 import { FundingCyclesFormProps, useFundingCyclesForm } from './hooks'
 
 const FundingCycleCallout: React.FC = () => {
