+++ src/components/Create/components/pages/NftRewards/NftRewardsPage.tsx
@@ -9,7 +9,7 @@ import { useAppSelector } from 'hooks/AppSelector'
 import { useModal } from 'hooks/Modal'
 import { useContext } from 'react'
 import { useSetCreateFurthestPageReached } from 'redux/hooks/EditingCreateFurthestPageReached'
-import { CreateButton } from '../../CreateButton'
+import { CreateButton } from 'components/CreateButton'
 import { CreateCollapse } from '../../CreateCollapse'
 import { OptionalHeader } from '../../OptionalHeader'
 import { RewardsList } from '../../RewardsList'
