+++ src/components/Create/components/pages/Payouts/PayoutsPage.tsx
@@ -4,11 +4,11 @@ import { Form, Space, Tooltip } from 'antd'
 import { useWatch } from 'antd/lib/form/Form'
 import { useContext, useEffect, useMemo } from 'react'
 import { useSetCreateFurthestPageReached } from 'redux/hooks/EditingCreateFurthestPageReached'
+import { allocationTotalPercentDoNotExceedTotalRule } from 'utils/antd-rules'
 import { Selection } from '../../Selection'
 import { SelectionCardProps } from '../../Selection/SelectionCard'
 import { Wizard } from '../../Wizard'
 import { PageContext } from '../../Wizard/contexts/PageContext'
-import { allocationTotalPercentDoNotExceedTotalRule } from '../utils'
 import { PayoutsList } from './components/PayoutsList'
 import { useAvailablePayoutsSelections, usePayoutsForm } from './hooks'
