+++ src/components/Create/components/pages/ProjectToken/components/CustomTokenSettings/CustomTokenSettings.tsx
@@ -10,9 +10,9 @@ import { TokenRedemptionRateGraph } from 'components/TokenRedemptionRateGraph'
 import { useAppSelector } from 'hooks/AppSelector'
 import useMobile from 'hooks/Mobile'
 import { useEditingDistributionLimit } from 'redux/hooks/EditingDistributionLimit'
+import { inputMustExistRule } from 'utils/antd-rules'
 import { formatAmount } from 'utils/formatAmount'
 import { MAX_DISTRIBUTION_LIMIT, MAX_MINT_RATE } from 'utils/v2v3/math'
-import { inputMustExistRule } from '../../../utils'
 import * as ProjectTokenForm from '../../hooks/ProjectTokenForm'
 import { ProjectTokensFormProps } from '../../hooks/ProjectTokenForm'
 import { ReservedTokenRateCallout, ReservedTokensList } from './components'
