+++ src/components/Create/components/pages/ProjectToken/hooks/ProjectTokenForm.ts
@@ -1,10 +1,7 @@
 import { Form } from 'antd'
 import { useWatch } from 'antd/lib/form/Form'
-import { AllocationSplit } from 'components/Create/components/Allocation'
-import {
-  allocationToSplit,
-  splitToAllocation,
-} from 'components/Create/utils/splitToAllocation'
+import { AllocationSplit } from 'components/Allocation'
+import { allocationToSplit, splitToAllocation } from 'utils/splitToAllocation'
 import { ONE_MILLION } from 'constants/numbers'
 import { useAppDispatch } from 'hooks/AppDispatch'
 import { useAppSelector } from 'hooks/AppSelector'
