+++ src/components/Create/components/pages/index.ts
@@ -6,5 +6,3 @@ export * from './ProjectDetails'
 export * from './ProjectToken'
 export * from './ReconfigurationRules'
 export * from './ReviewDeploy'
-// TODO: We should move these somewhere higher up so all can use
-export * from './utils'
