--- src/components/Create/components/pages/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './rules'
