+++ src/components/PayoutCard/PayoutCard.tsx
@@ -10,8 +10,8 @@ import { useCallback } from 'react'
 import { stopPropagation } from 'react-stop-propagation'
 import { classNames } from 'utils/classNames'
 import { formatDate } from 'utils/format/formatDate'
-import { Allocation, AllocationSplit } from '../Allocation'
-import { DeleteConfirmationModal } from '../DeleteConfirmationModal'
+import { Allocation, AllocationSplit } from 'components/Allocation'
+import { DeleteConfirmationModal } from 'components/DeleteConfirmationModal'
 import { Amount } from './Amount'
 import { isProjectSplit } from 'utils/splits'
