+++ src/utils/antd-rules/index.ts
@@ -7,3 +7,4 @@ export * from './inputIsValidUrlRule'
 export * from './inputMustBeEthAddressRule'
 export * from './inputMustExistRule'
 export * from './inputNonZeroRule'
+export * from './currencyAmountMustExistRule'
