+++ src/utils/format/formatFundingTarget.ts
@@ -5,6 +5,9 @@ import { formatCurrencyAmount } from 'utils/formatCurrencyAmount'
 import { V2V3_CURRENCY_ETH } from 'utils/v2v3/currency'
 import { MAX_DISTRIBUTION_LIMIT } from 'utils/v2v3/math'
 
+// Formats a funding target into the format:
+// {prefix} 99,999 {suffix}
+// where prefix and suffix is set based on distributionLimitCurrency
 export function formatFundingTarget({
   distributionLimitWad,
   distributionLimitCurrency = V2V3_CURRENCY_ETH,
