+++ src/utils/splitToAllocation.ts
@@ -1,7 +1,7 @@
 import { defaultSplit, Split } from 'models/splits'
 import { sanitizeSplit } from 'utils/splits'
 import { preciseFormatSplitPercent, splitPercentFrom } from 'utils/v2v3/math'
-import { AllocationSplit } from '../components/Allocation'
+import { AllocationSplit } from 'components/Allocation'
 
 export const splitToAllocation = (split: Split): AllocationSplit => {
   return {
